Try to get cookies from backend, set env var
This does the setup to get login and CSRF cookies from the backend.

There was great frustration with trying to set a custom environment
variable. The key is that environment variables passed in to the web app
need to be prepended with REACT_APP_. The running web app in the browser
does not have access to the process environment. The access is simulated
by something like Webpack doing a textual find and replace to inject
environment variables at compile time.

https://create-react-app.dev/docs/adding-custom-environment-variables/